backend/requirements.txt
fastapi
sqlalchemy
python-dotenv
python-multipart
uvicorn
numpy
groq
pydantic
requests
librosa
pydub 
scipy